fix: deep_copy model 로 옮김
